requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx[http2]>=0.27.0
h2>=4.1.0
protobuf>=5.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0
